backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
# PaddleOCR - install separately for Windows compatibility
# paddlepaddle - install from: https://www.paddlepaddle.org.cn/install/quick?docurl=/documentation/docs/zh/install/pip/windows-pip.html
paddleocr>=2.7.0
pillow==10.1.0
numpy==1.24.3
transformers==4.35.0
torch==2.1.0
pydantic==2.5.0
accelerate==0.24.1
sentencepiece==0.1.99
protobuf==3.20.3
# Alternative OCR for Windows fallback
easyocr>=1.7.0
# PDF text extraction
PyPDF2>=3.0.0

